Moved timestamp into own protocol, updated tests
