+++ GRDBKit/Model.swift
@@ -14,23 +14,6 @@ public protocol GRDBModel {
     associatedtype ID = GRDBKit.ID
     typealias IDKey = WritableKeyPath<Self, ID?>
     static var idKey: IDKey { get }
-    
-    // Timestamps
-    typealias TimestampKey = WritableKeyPath<Self, Date?>
-    static var createdAtKey: TimestampKey? { get }
-    static var updatedAtKey: TimestampKey? { get }
-}
-
-extension GRDBModel {
-    /// See `Model`.
-    public static var createdAtKey: TimestampKey? {
-        return nil
-    }
-    
-    /// See `Model`.
-    public static var updatedAtKey: TimestampKey? {
-        return nil
-    }
 }
 
 extension GRDBModel {
@@ -52,59 +35,6 @@ extension GRDBModel {
         set {
             let path = Self.idKey
             self[keyPath: path] = newValue
-        }
-    }
-    
-    /// Access the timestamp keyed by `createdAtKey`.
-    public var modelCreatedAt: Date? {
-        get {
-            guard let createdAt = Self.createdAtKey else {
-                return nil
-            }
-            return self[keyPath: createdAt]
-        }
-        set {
-            guard let createdAt = Self.createdAtKey else {
-                return
-            }
-            self[keyPath: createdAt] = newValue
-        }
-    }
-    
-    /// Access the timestamp keyed by `updatedAtKey`.
-    public var modelUpdatedAt: Date? {
-        get {
-            guard let updatedAt = Self.updatedAtKey else {
-                return nil
-            }
-            return self[keyPath: updatedAt]
-        }
-        set {
-            guard let updatedAt = Self.updatedAtKey else {
-                return
-            }
-            self[keyPath: updatedAt] = newValue
-        }
-    }
-}
-
-extension GRDBModel where Self: MutablePersistableRecord {
-    mutating func update(timestamping db: Database, date: Date = .init()) throws {
-        modelUpdatedAt = date
-        try update(db)
-    }
-    
-    mutating func insert(timestamping db: Database, date: Date = .init()) throws {
-        modelCreatedAt = date
-        modelUpdatedAt = date
-        try insert(db)
-    }
-    
-    mutating func save(timestamping db: Database, date: Date = .init()) throws {
-        do {
-            try update(timestamping: db, date: date)
-        } catch PersistenceError.recordNotFound {
-            try insert(timestamping: db, date: date)
         }
     }
 }
